Now searches for the required fields instead of relying on line numbers.
The line number system used in the previous version was for proof of concept. Given that it is expected that different YAML files will be of different lengths, this version removes the reliance on line numbers and instead searches through the file for the fields that are to be checked.